requirements.txt
qutip
numpy
matplotlib
scipy
tqdm
pyyaml
